requirements.txt
pycryptodome

ttkbootstrap
